Move year quarter classses to time package
