+++ src/custom-functions/bcode-quarter.ts
@@ -2,7 +2,7 @@ import { ApiResponseError } from './error'
 import { CachingBuffettCodeApiClientV2 } from '../api/caching-client'
 import { CachingQuarterProperty } from '../api/caching-quarter-property'
 import { Result } from '../result'
-import { YearQuarter } from '../year-quarter'
+import { YearQuarter } from '../time/year-quarter'
 
 export function bcodeQuarter(
   client: CachingBuffettCodeApiClientV2,
