+++ src/services/quarter-cache.ts
@@ -1,4 +1,4 @@
-import { YearQuarter } from '../year-quarter'
+import { YearQuarter } from '../time/year-quarter'
 
 export class QuarterCache {
   static readonly prefix = 'quarter'
